requirements.txt
pandas
numpy
matplotlib
seaborn
biopython
openpyxl
matplotlib-venn
